pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cad-agents"
version = "0.1.0"
description = "A system of agents designed to help engineers with CAD modeling tasks using Google's ADK"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Samuel Larson", email = "your.email@example.com" }
]
dependencies = [
    "google-adk>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["cad_agents"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"] 